订单关系，publish，还有实体类加主键注解

+++ miscroservice_user/src/main/java/com/e_commerce/miscroservice/user/dao/UserTimeRecordDao.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface UserTimeRecordDao {
     List<TUserTimeRecord> selectMonthlyTimeRecord(Long userId, Long begin, Long end);
 
-    int insert(TUserTimeRecord record);
+    long insert(TUserTimeRecord record);
 }
