+++ miscroservice_user/src/main/java/com/e_commerce/miscroservice/user/dao/impl/UserTimeRecordDaoImpl.java
@@ -17,7 +17,7 @@ public class UserTimeRecordDaoImpl implements UserTimeRecordDao {
     }
 
     @Override
-    public int insert(TUserTimeRecord record) {
+    public long insert(TUserTimeRecord record) {
         return MybatisOperaterUtil.getInstance().save(record);
     }
 }
